refactor: add gql queries file & move interface
